User management updates
The Users page now loads a sidebar to edit their name
or email.

Admins can register a user for an event directly. This
allows an admin to register someone who missed a
registration deadline but still attended the event.

Users cannot be removed from events because we want
to track registrations compared with attendance to find
paterns of behavior.